Fix an issue mostly caused by a bad test setup, although the size of state.captured was wrong too

+++ src/state.h
@@ -69,7 +69,7 @@ struct State {
     /* Derived information */
 
     // What squares each player has captured
-    uint_fast8_t captured[NUM_PLAYERS];
+    uint_fast16_t captured[NUM_PLAYERS];
     uint_fast64_t blocked[NUM_PLAYERS];
 
     uint_fast8_t largestNetworkSize;
